Added SmallestIntegralType helper template, which is used to derive FixedPoint<traits, bits, power>::QNumberType.
QNumberType represents an integer big enough to hold a FixedPoint number in raw Qm.n fixed-point format.

QNumberType values are returned from the getQNumber() member and also via the implicit QNumberType conversion operator.

The underlying conversion, which needs to be verified, shifts the data 'bits' by 'power' iff 'power' > 0.
This will give the receiver a reduced-resolution integer of the correct rank, where the resolution is 'bits' and the rank is 'bits' + 'power'.

For zero and negative 'power' values, the data 'bits' are returned verbatim.
It is understood that the receiver knows the underlying Qm.n format and is aware that returned integer represents a potentially fractional value.