Added bitwise-or tests.  Unfortunately the tests do not build. This is due to an ambiguous operator overload. Basically speaking, the compiler can't tell whether to use the OE overload or the native variant. This is because native integral types are implicitly convertible to OE equivalents via an OE constructor, and OE objects are implicitly convertible to the underlying integral type via implicit type conversion.
At this stage I want to keep both implicit conversions, but this means heaps of boiler plate to
create all the perturbations of operator overloads to force precedence.

However, there seems to be a nice wee Github project specifically for this...

https://github.com/taocpp/operators

It's MIT licensed, so I'll check it out next time.
My only concern is the introduction of code bloat,
but the compiler should be able to optimize all this syntactic noise away.

We'll see what happens next time...